feat: add bg locale translations (#812)
The translations for days, months, and units in the Bulgarian (bg) locale have been updated. This update includes changes to the abbreviated, narrow, short, and wide forms of the translations. The translations for relative time expressions (e.g., "след" and "преди") have also been updated for the year, month, week, and day units.