Don't eagerly load the Locale, nor pytest until it's needed.
The main problem here was that importing pendulum.testing.traveller was
importing time_machine, which was also importing pytest.

And creating the DifferenceFormatter at import time was also loading the
locale.

Both of these are "optional" in the sense that there is lots you can do in
Pendulum without ever using them, so we can relatively straightforwardly delay
things until they are requested.

Tested by running `uv run --with time_machine -p 3.12 python -Ximporttime -c 'import pendulum'`:

- Before: ~50-64ms (a lot of noise)
- After: 20-29ms

Not a huge slow down, but I figured it was worth it anyway, especially not not
load pytest.